Modernize the `DocumentStatusPill` component
- convert from classic to glimmer component
- convert to .gjs
- remove css hacks
