+++ app/components/bbcdr/report-edit-header.hbs
@@ -21,7 +21,7 @@
     </p>
     {{/unless}}
     <p>
-      {{shared/document-status-pill @report.status}}
+      <Bbcdr::DocumentStatus @status={{@report.status}} />
     </p>
   </AuContent>
 </div>
