--- app/components/shared/document-status-pill.hbs
@@ -1,1 +1,0 @@
-{{capitalize this.status.label}}
