+++ app/styles/app.scss
@@ -21,7 +21,6 @@
 @import "@appuniversum/ember-appuniversum/styles/a-components";
 @import "project/c-decision-articles-field";
 @import "project/c-data-list";
-@import "project/c-document-status-pills";
 @import "project/c-sidebar-container";
 @import "project/c-status-label";
