--- app/styles/project/_c-document-status-pills.scss
@@ -1,24 +1,0 @@
-/* ==================================
-   #DOCUMENT STATUS PILLS
-   Maps the status pills on resource tags
-   ================================== */
-
-.au-c-pill[resource="http://data.lblod.info/document-statuses/concept"] {
-  @extend .au-c-pill--border;
-}
-
-.au-c-pill[resource="http://data.lblod.info/document-statuses/goedgekeurd"] {
-  @extend .au-c-pill--success;
-}
-
-.au-c-pill[resource="http://data.lblod.info/document-statuses/opmerking"] {
-  @extend .au-c-pill--warning;
-}
-
-.au-c-pill[resource="http://data.lblod.info/document-statuses/verstuurd"] {
-  @extend .au-c-pill--action !optional;
-}
-
-.au-c-pill[resource="http://data.lblod.info/document-statuses/prullenbak"] {
-  @extend .au-c-pill--default;
-}
